requirements.txt
python-dotenv==1.0.1
lxml==5.1.0
beautifulsoup4==4.12.3
requests==2.31.0
django_filter==23.5
Markdown==3.5.2
djangorestframework==3.14.0
Django==5.0.2